requirements.txt
e3nn~=0.3.5
filelock~=3.1.0
horovod~=0.22.1
pytorch-lightning~=1.4.8

-f https://pytorch-geometric.com/whl/torch-1.9.0+cu111.html
torch-scatter~=2.0.8
torch-sparse~=0.6.12
torch-cluster~=1.5.9
torch-spline-conv~=1.2.1
torch-geometric~=2.0.1